# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Movies DB
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  #MOVIES
  /movies:
    # binds a127 app logic to a route
    x-swagger-router-controller: movies
    get:
      description: Returns all movies from database
      # used as the method name of the controller
      operationId: movies
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MoviesResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /movie/{idMovie}:
    # binds a127 app logic to a route
    x-swagger-router-controller: movies
    get:
      description: Returns one movie by ID
      # used as the method name of the controller
      operationId: moviebyID
      parameters:
        - name: idMovie
          in: path
          description: ID
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MovieResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /movie/save:
    # binds a127 app logic to a route
    x-swagger-router-controller: movies
    post:
      description: Save a new movie
      # used as the method name of the controller
      operationId: saveMovie
      parameters:
        - name: data
          in: body
          description: body form
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MovieResponseSave"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /movie/update/{idMovie}:
    # binds a127 app logic to a route
    x-swagger-router-controller: movies
    put:
      description: Update a movie
      # used as the method name of the controller
      operationId: updateMovie
      parameters:
        - name: idMovie
          in: path
          description: body form
          required: true
          type: string
        - name: data
          in: body
          description: body form
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MovieResponseSave"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /movie/delete/{idMovie}:
    # binds a127 app logic to a route
    x-swagger-router-controller: movies
    delete:
      description: Delete a movie
      # used as the method name of the controller
      operationId: deleteMovie
      parameters:
        - name: idMovie
          in: path
          description: id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MovieResponseSave"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  #GENRES
  /genres:
    # binds a127 app logic to a route
    x-swagger-router-controller: genres
    get:
      description: Returns all genres from database
      # used as the method name of the controller
      operationId: genres
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GenresResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /genre/{idGenre}:
    # binds a127 app logic to a route
    x-swagger-router-controller: genres
    get:
      description: Returns one genre by ID
      # used as the method name of the controller
      operationId: genrebyID
      parameters:
        - name: idGenre
          in: path
          description: ID
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GenreResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /genre/save:
    # binds a127 app logic to a route
    x-swagger-router-controller: genres
    post:
      description: Save a genre movie
      # used as the method name of the controller
      operationId: saveGenre
      parameters:
        - name: data
          in: body
          description: body form
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GenreResponseSave"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /genre/update/{idGenre}:
    # binds a127 app logic to a route
    x-swagger-router-controller: genres
    put:
      description: Update a genre
      # used as the method name of the controller
      operationId: updateGenre
      parameters:
        - name: idGenre
          in: path
          description: body form
          required: true
          type: string
        - name: data
          in: body
          description: body form
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GenreResponseSave"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /genre/delete/{idGenre}:
    # binds a127 app logic to a route
    x-swagger-router-controller: genres
    delete:
      description: Delete a genre
      # used as the method name of the controller
      operationId: deleteGenre
      parameters:
        - name: idGenre
          in: path
          description: id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GenreResponseSave"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  MoviesResponse:
    required:
      - message
    properties:
      message:
        type: string
  MovieResponseSave:
    required:
      - message
    properties:
      message:
        type: string
  MovieResponse:
      required:
        - _id
        - title
        - director
        - studio
        - mppa_rating
        - status
      properties:
        _id:
          type: string
        title:
          type: string
        director:
          type: string
        studio:
          type: string
        mppa_rating:
          type: string
        status:
          type: string
  GenresResponse:
    required:
      - message
    properties:
      message:
        type: string
  GenreResponseSave:
    required:
      - message
    properties:
      message:
        type: string
  GenreResponse:
      required:
        - _id
        - name
        - status
      properties:
        _id:
          type: string
        name:
          type: string
        status:
          type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string